changed the smallest word dimension from 64 bits to 32 bits
svn path=/trunk/iCub/contrib/src/eMorph/; revision=15020

+++ src/aexGrabber/include/iCub/device2yarp.h
@@ -59,7 +59,7 @@ struct aer {
 };
 
 struct atom {
-    u16 data;
+    u32 data;
 };
